remove enabled flag from merkle tree proofs
